Remove temp file only if it exists

+++ credentials/remove-newlines.sh
@@ -4,5 +4,5 @@ for f in "$(dirname "$0")/"*".txt"; do
 	cat "$f" | tr -d '\n' > "$TMPFILE"
 	mv "$TMPFILE" "$f"
 done
-rm "$TMPFILE"
+[ -e "$TMPFILE" ] && rm "$TMPFILE"
